fuzion_rt: change fzE_last_error to i64
windows GetLastError returns DWORD which is unsigned. So changing this to i64 to be able to contain this.

+++ include/fz.h
@@ -46,7 +46,7 @@ void fzE_memcpy(void *restrict dest, const void *restrict src, size_t sz);
 
 // returns the latest error number of
 // the current thread
-int fzE_last_error(void);
+int64_t fzE_last_error(void);
 
 // NYI: UNDER DEVELOPMENT: fzE_last_error_as_string, returning the error as a human readable string
